+++ gframe/generic_duel.h
@@ -85,12 +85,12 @@ protected:
 	bool CheckReady();
 	uint32_t GetCount(const std::vector<duelist>& players);
 	bool CheckFree(const std::vector<duelist>& players);
-	int GetFirstFree(int start = 0);
+	uint8_t GetFirstFree(uint8_t start = 0);
 	void SetAtPos(DuelPlayer* dp, size_t pos);
 	duelist& GetAtPos(uint8_t pos);
 	void Catchup(DuelPlayer* dp);
-	int GetPos(DuelPlayer* dp);
-	void OrderPlayers(std::vector<duelist>& players, int offset = 0);
+	uint8_t GetPos(DuelPlayer* dp);
+	void OrderPlayers(std::vector<duelist>& players, size_t offset = 0);
 	template<typename T, typename T2>
 	inline void Iter(T& list, const T2& func) {
 		for(auto& dlr : list) {
