+++ gframe/repo_manager.cpp
@@ -372,7 +372,7 @@ void RepoManager::CheckoutCb(const char* path, size_t completed_steps, size_t to
 		percent = CHECKOUT_PERCENTAGE;
 	else {
 		static constexpr auto DELTA_INCREMENT = CHECKOUT_PERCENTAGE - DELTA_OBJECTS_PERCENTAGE;
-		percent = DELTA_OBJECTS_PERCENTAGE + ((DELTA_INCREMENT * completed_steps) / total_steps);
+		percent = static_cast<int>(DELTA_OBJECTS_PERCENTAGE + ((DELTA_INCREMENT * completed_steps) / total_steps));
 	}
 	auto pl = static_cast<GitCbPayload*>(payload);
 	pl->rm->SetRepoPercentage(pl->path, percent);
